Revert "类型转换警告处理"
This reverts commit 926c930f2e9ef1dba05573459014987fc674efa9.

+++ src/Camera.cpp
@@ -28,7 +28,7 @@ int Camera::getOffsetX()
     {
         tx = WORLD_RIGHT - GAME_WIDTH;
     }
-    return static_cast<int>(tx);
+    return tx;
 } 
 
 int Camera::getOffsetY()
